+++ src/main/java/com/accenture/flowershop/be/entity/order/Item.java
@@ -2,8 +2,10 @@ package com.accenture.flowershop.be.entity.order;
 
 import com.accenture.flowershop.be.entity.flower.Flower;
 
+import java.math.BigDecimal;
+
 public class Item {
     private Flower flower;
     private int amount;
-    private double subTotal;
+    private BigDecimal subTotal;
 }
